pyproject.toml
[project]
name = "adk-agent-sim"
version = "0.1.0"
description = "A simulator that allows a human to perform the task of an ADK agent"
authors = [
    { name = "Kevin Aud" }
]
requires-python = ">=3.14"
dependencies = [
  "python-dotenv>=1.0.1",
  "typer>=0.12.5",
  "rich>=13.8.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.12.0",
  "structlog>=25.5.0",
  "tenacity>=9.1.2",
  "httpx>=0.28.1",
  "google-adk>=1.21.0",
  "fastmcp>=2.14.0",
  "nicegui>=2.0.0",
]

[project.scripts]
demo = "adk_agent_sim.demo.demo:app"
demo-server = "adk_agent_sim.demo.mcp_servers.demo_server:run_server"

[tool.setuptools]
packages = ["adk_agent_sim"]

[tool.uv]
package = true

[tool.uv.sources]
specify-cli = { git = "https://github.com/github/spec-kit.git" }

[tool.ruff]
line-length = 88
indent-width = 2

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportPrivateImportUsage = false
reportMissingTypeStubs = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportMissingTypeArgument = false
reportAttributeAccessIssue = false
reportPrivateUsage = false
reportRedeclaration = false
reportUnusedClass = false

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require API keys, slower)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pyright>=1.1.407",
    "ruff>=0.14.7",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "specify-cli",
]
